refactored to stronghold package
