+++ build.properties
@@ -2,7 +2,7 @@
 
 
 # Project specific information
-package=com.nutrons.steamworks
+package=com.nutrons.stronghold
 robot.class=${package}.RobotBootstrapper
 test.dir=test
 simulation.world.file=/usr/share/frcsim/worlds/GearsBotDemo.world
