+++ src/com/nutrons/stronghold/RobotBootstrapper.java
@@ -1,4 +1,4 @@
-package com.nutrons.steamworks;
+package com.nutrons.stronghold;
 
 import com.nutrons.framework.Robot;
 import com.nutrons.framework.StreamManager;
